vecbench/requirements.txt
pgvector==0.2.5
SQLAlchemy==2.0.28
psycopg[binary]==3.1.18
psycopg2-binary==2.9.9
influxdb-client==1.41.0
deepdish==0.3.7
apscheduler==3.10.4
pyaml-env==1.2.1
ray[default]==2.10.0
google-api-python-client==2.123.0
cryptography==42.0.5
google-cloud-compute==1.18.0
google-cloud-storage==2.16.0
tqdm==4.66.3
pandas==2.0.3
numpy==1.23.4
opentelemetry-api==1.14.0
opentelemetry-sdk==1.14.0
opentelemetry-exporter-gcp-monitoring==1.6.0a0
opentelemetry-propagator-gcp==1.6.0
tinyflux==0.4.1
py-spy==0.3.14
Jinja2==3.1.3
google-cloud-aiplatform==1.44.0
google-auth==2.29.0
grpcio==1.58.0
pyarrow==15.0.2
pinecone-client[grpc]==3.2.0
redis==5.0.3
sqlalchemy-spanner==1.6.2
pytest==8.1.1
google-api-core==2.18.0
mysql-connector-python==8.3.0
pymilvus==2.3.4
